% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grammar.R
\name{check_grammar}
\alias{check_grammar}
\title{check_grammar}
\usage{
check_grammar(text, options = list())
}
\arguments{
\item{text}{The text to check}

\item{options}{Which rules should be ignored see https://github.com/btford/write-good for details.}
}
\value{
A character vector of suggestions for grammar changes.
}
\description{
Checks a piece of text for common grammatical errors and returns a list of
suggestions.
}
\examples{

check_grammar("So the cat was stolen.", list(passive = FALSE))


}
