% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grammar.R
\name{check_project}
\alias{check_project}
\title{Check All Files in a Project}
\usage{
check_project(path = ".", exclude_chunks = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{the path of the project}

\item{exclude_chunks}{exclude RMarkdown code chunks. Defaults to FALSE.}

\item{verbose}{display a message in case of good writing. Defaults to TRUE.}
}
\value{
a data.frame with all suggestions
}
\description{
Check all md and Rmd files in a project.
}
