% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gramr.R
\docType{package}
\name{gramr}
\alias{gramr}
\alias{gramr-package}
\title{A R package To Check for Grammar within RMarkdown documents.}
\description{
Provides additional checks in grammar not currently caught
by the native spell check within RStudio.
}
