% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rmd.R
\name{parse_rmd}
\alias{parse_rmd}
\title{Parse RMarkdown documents}
\usage{
parse_rmd(file)
}
\arguments{
\item{file}{a character vector of the file or path}
}
\value{
used in the shiy app
}
\description{
Parse RMarkdown documents
}
\examples{
# don't run
# parse_rmd()
}
