% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grammar.R
\name{write_good_file}
\alias{write_good_file}
\title{Check the grammar of a Rmd by filename}
\usage{
write_good_file(filename = "")
}
\arguments{
\item{filename}{the name of an Rmd file. Does not have to be open in RStudio.}
}
\value{
a data.frame with suggestions for grammar fixes
}
\description{
Check the grammar of a Rmd by filename
}
\examples{
# don't run during tests
# write_good_file()
}
