% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grammar.R
\name{write_good_ip}
\alias{write_good_ip}
\title{Check the grammar of an in-progress document within RStudio
If you have an 'UntitledX' document check the grammar before saving the file.}
\usage{
write_good_ip()
}
\value{
a print out of suggestions for grammar fixes
}
\description{
Check the grammar of an in-progress document within RStudio
If you have an 'UntitledX' document check the grammar before saving the file.
}
\examples{
# don't run during tests
# write_good_ip()
}
